move unnecessary methods out of thunk (#1246)
