+++ runtime/thunk/thunk.ts
@@ -2,10 +2,8 @@ import {AuthenticationType} from '../../types';
 import type {ExecutionContext} from '../../api_types';
 import type {FetchRequest} from '../../api_types';
 import type {FetchResponse} from '../../api_types';
-import type {Formula} from '../../api';
 import type {FormulaSpecification} from '../types';
 import {FormulaType} from '../types';
-import type {GenericSyncFormula} from '../../api';
 import type {MetadataFormula} from '../../api';
 import {MetadataFormulaType} from '../types';
 import type {PackFormulaResult} from '../../api_types';
@@ -20,58 +18,14 @@ import type {SyncExecutionContext} from '../../api_types';
 import type {SyncFormulaResult} from '../../api';
 import type {SyncFormulaSpecification} from '../types';
 import type {TypedPackFormula} from '../../api';
+import {findFormula} from '../common/helpers';
+import {findSyncFormula} from '../common/helpers';
 import {isDynamicSyncTable} from '../../api';
 import {unwrapError} from '../common/marshaling';
 import {wrapError} from '../common/marshaling';
 
 export {marshalValue} from '../common/marshaling';
 export {unmarshalValue} from '../common/marshaling';
-
-export function findFormula(packDef: PackVersionDefinition, formulaNameWithNamespace: string): Formula {
-  const packFormulas = packDef.formulas;
-  if (!packFormulas) {
-    throw new Error(`Pack definition has no formulas.`);
-  }
-
-  const [namespace, name] = formulaNameWithNamespace.includes('::')
-    ? formulaNameWithNamespace.split('::')
-    : ['', formulaNameWithNamespace];
-
-  if (namespace) {
-    // eslint-disable-next-line no-console
-    console.log(
-      `Warning: formula was invoked with a namespace (${formulaNameWithNamespace}), but namespaces are now deprecated.`,
-    );
-  }
-
-  const formulas: Formula[] = Array.isArray(packFormulas) ? packFormulas : packFormulas[namespace];
-  if (!formulas || !formulas.length) {
-    throw new Error(`Pack definition has no formulas${namespace ?? ` for namespace "${namespace}"`}.`);
-  }
-  for (const formula of formulas) {
-    if (formula.name === name) {
-      return formula;
-    }
-  }
-  throw new Error(`Pack definition has no formula "${name}"${namespace ?? ` in namespace "${namespace}"`}.`);
-}
-
-export function findSyncFormula(packDef: PackVersionDefinition, syncFormulaName: string): GenericSyncFormula {
-  if (!packDef.syncTables) {
-    throw new Error(`Pack definition has no sync tables.`);
-  }
-
-  for (const syncTable of packDef.syncTables) {
-    const syncFormula = syncTable.getter;
-    // TODO(huayang): it seems like a bug that client passes syncTable.name in, instead of syncTable.identityName or
-    // syncTable.getter.name.
-    if (syncTable.name === syncFormulaName) {
-      return syncFormula;
-    }
-  }
-
-  throw new Error(`Pack definition has no sync formula "${syncFormulaName}" in its sync tables.`);
-}
 
 /**
  * The thunk entrypoint - the first code that runs inside the v8 isolate once control is passed over.
@@ -89,21 +43,6 @@ export async function findAndExecutePackFunction<T extends FormulaSpecification>
     // all errors should be marshaled to avoid IVM dropping essential fields / name.
     throw shouldWrapError ? wrapError(err) : err;
   }
-}
-
-export function tryFindFormula(packDef: PackVersionDefinition, formulaNameWithNamespace: string): Formula | undefined {
-  try {
-    return findFormula(packDef, formulaNameWithNamespace);
-  } catch (_err) {}
-}
-
-export function tryFindSyncFormula(
-  packDef: PackVersionDefinition,
-  syncFormulaName: string,
-): GenericSyncFormula | undefined {
-  try {
-    return findSyncFormula(packDef, syncFormulaName);
-  } catch (_err) {}
 }
 
 function doFindAndExecutePackFunction<T extends FormulaSpecification>(
